requirements.txt
mcp
python-dotenv
anyio